requirements.txt
flask
flask-cors
openai
requests
cloudinary
moviepy
pillow
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
